### 2023-07-23 20dc3e0
Update paper.bib
Fix DOI for Paper

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
