+++ paper/paper.bib
@@ -15,7 +15,7 @@ volume = {147},
 pages = {107828},
 year = {2023},
 issn = {0747-5632},
-doi = {https://doi.org/10.1016/j.chb.2023.107828},
+doi = {10.1016/j.chb.2023.107828},
 url = {https://www.sciencedirect.com/science/article/pii/S0747563223001796},
 author = {Lena Fanya Aeschbach and Dominik Kayser and Antony {Berbert De Castro Hüsler} and Klaus Opwis and Florian Brühlmann},
 keywords = {Overestimation, , Esports, Motivated reasoning, Dunning–Kruger effect},
