### 2023-07-23 b424699
Update paper.md
Add affiliation.

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
