+++ pyproject.toml
@@ -87,7 +87,7 @@ docs = [
     "sphinx-autoapi~=2.1.1",
     "sphinxext-opengraph~=0.8",
     "sphinxcontrib-bibtex~=2.5.0",
-    "sphinx-autodoc-typehints~=1.23.3",
+    "sphinx-autodoc-typehints~=1.24.0",
 ]
 tests = [
     "tox~=4.6",
