### 2023-08-01 #5682 @vivekjoshy
Some potential reviewers:

- [matt-graham](https://github.com/matt-graham)
- [dufourc1](https://github.com/dufourc1)
- [gvieralopez](https://github.com/gvieralopez)
- [Naeemkh](https://github.com/Naeemkh)
- [olgadoronina](https://github.com/olgadoronina)

Couldn't find any reviewers that mentioned expertise in Ranking Systems like Plackett-Luce, Bradley-Terry or Glicko.
