### 2023-08-06 b7b4918
Update CODE_OF_CONDUCT.md
Adopt project's code of conduct.

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
