+++ CODE_OF_CONDUCT.md
@@ -1,94 +1,49 @@
-# Citizen Code of Conduct
+# Code of Conduct - Open Debates Project
 
-## 1. Purpose
+The Open Debates Project is an open-source project that aims to develop mathematical models for the digital verification of argument soundness. To foster a positive and inclusive community, we have established this Code of Conduct, which applies to all projects within the scope of the Open Debates Project. All participants are expected to adhere to these guidelines in order to maintain a friendly and respectful environment.
 
-A primary goal of OpenSkill Python is to be inclusive to the largest number of contributors, with the most varied and diverse backgrounds possible. As such, we are committed to providing a friendly, safe and welcoming environment for all, regardless of gender, sexual orientation, ability, ethnicity, socioeconomic status, and religion (or lack thereof).
+## Reporting Grievances
 
-This code of conduct outlines our expectations for all those who participate in our community, as well as the consequences for unacceptable behavior.
+If you have any grievances related to any project within the Open Debates Project, such as accusations or concerns about inappropriate behavior, please address them in a private manner. The appropriate channel for reporting grievances is by contacting the Steering Council or the Benevolent Dictator for Life (BDFL) privately. 
 
-We invite all those who participate in OpenSkill Python to help us create safe and positive experiences for everyone.
+Please do not publicize details of your grievances until after an investigation has concluded. Publicizing grievances prematurely will result in voiding impeachments and the barring of accused members from the community. Additionally, it is important to provide evidence to support any claims of inappropriate behavior to ensure fair and objective evaluations.
 
-## 2. Open [Source/Culture/Tech] Citizenship
+You can contact the BDFL via email at vivek@opendebates.net for reporting any grievances.
 
-A supplemental goal of this Code of Conduct is to increase open [source/culture/tech] citizenship by encouraging participants to recognize and strengthen the relationships between our actions and their effects on our community.
+## Compliance with Laws
 
-Communities mirror the societies in which they exist and positive action is essential to counteract the many forms of inequality and abuses of power that exist in society.
+All participants in any project within the Open Debates Project are expected to comply with the laws of their respective countries while engaging in these projects. Violations of legal obligations will not be tolerated.
 
-If you see someone who is making an extra effort to ensure our community is welcoming, friendly, and encourages all participants to contribute to the fullest extent, we want to know.
+## Appropriateness in Communication
 
-## 3. Expected Behavior
+To ensure a professional and inclusive atmosphere, it is essential that all communication within the projects adheres to certain standards:
 
-The following behaviors are expected and requested of all community members:
+1. **Inappropriate Content**: Do not portray or send inappropriate content in any public channels or areas of the projects. This includes but is not limited to offensive language, explicit material, or discriminatory content.
 
- * Participate in an authentic and active way. In doing so, you contribute to the health and longevity of this community.
- * Exercise consideration and respect in your speech and actions.
- * Attempt collaboration before conflict.
- * Refrain from demeaning, discriminatory, or harassing behavior and speech.
- * Be mindful of your surroundings and of your fellow participants. Alert community leaders if you notice a dangerous situation, someone in distress, or violations of this Code of Conduct, even if they seem inconsequential.
- * Remember that community event venues may be shared with members of the public; please be respectful to all patrons of these locations.
+2. **Politics**: Political content should not be discussed in public venues within the projects. Our focus is solely on the development and advancement of mathematical models for argument verification.
 
-## 4. Unacceptable Behavior
+## Active Participation
 
-The following behaviors are considered harassment and are unacceptable within our community:
[…]
