### 2023-08-11 aa59468
Add Algolia Support
Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
