### 2023-08-11 7e2f977
Add Algolia Support (#105)
* Add Algolia Support

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Add Changelog Fragment

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

---------

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
