+++ benchmark/processors/rank.py
@@ -1,17 +1,13 @@
-import itertools
-import math
 import time
 from typing import Union
 
 import jsonlines
 import numpy as np
-import trueskill
 from prompt_toolkit import HTML
 from prompt_toolkit import print_formatted_text as print
 from prompt_toolkit.shortcuts import ProgressBar
 from sklearn.model_selection import train_test_split
 
-import openskill
 from openskill.models import (
     BradleyTerryFull,
     BradleyTerryPart,
@@ -21,20 +17,12 @@ from openskill.models import (
 )
 
 
-def win_probability(team1, team2):
-    delta_mu = sum(r.mu for r in team1) - sum(r.mu for r in team2)
-    sum_sigma = sum(r.sigma**2 for r in itertools.chain(team1, team2))
-    size = len(team1) + len(team2)
-    denom = math.sqrt(size * (trueskill.BETA * trueskill.BETA) + sum_sigma)
-    ts = trueskill.global_env()
-    return ts.cdf(delta_mu / denom)
-
-
 class Rank:
     def __init__(
         self,
         path,
         seed: int,
+        minimum_matches: int,
         model: Union[
             BradleyTerryFull,
             BradleyTerryPart,
@@ -45,15 +33,15 @@ class Rank:
     ):
         self.data = list(jsonlines.open(path).iter())
         self.seed = seed
+        self.minimum_matches = minimum_matches
         self.model = model
 
         # Counters
         self.match_count = {}
-        self.confident_matches = 0
+        self.available_matches = 0
+        self.valid_matches = 0
         self.openskill_correct_predictions = 0
         self.openskill_incorrect_predictions = 0
-        self.trueskill_correct_predictions = 0
-        self.trueskill_incorrect_predictions = 0
 
         # Post Verification of Data
         self.verified_matches = []
@@ -65,11 +53,9 @@ class Rank:
 
         # Ratings
         self.openskill_players = {}
-        self.trueskill_players = {}
 
         # Time Spent
         self.openskill_time = None
-        self.trueskill_time = None
 
     def process(self):
         title = HTML(f'<style fg="Red">Counting Matches</style>')
@@ -103,23 +89,13 @@ class Rank:
         os_process_time_stop = time.time()
         self.openskill_time = os_process_time_stop - os_process_time_start
 
-        # Process TrueSkill Ratings
-        title = HTML(
-            f'Updating Ratings with <style fg="Green">TrueSkill</style> Model:'
-        )
-        with ProgressBar(title=title) as progress_bar:
-            ts_process_time_start = time.time()
-            for match in progress_bar(self.training_set, total=len(self.training_set)):
-                self.process_trueskill(match=match)
-        ts_process_time_stop = time.time()
-        self.trueskill_time = ts_process_time_stop - ts_process_time_start
-
         # Process Test Set
         title = HTML(f'<style fg="Red">Processing Test Set</style>')
         with ProgressBar(title=title) as progress_bar:
             for match in progress_bar(self.test_set, total=len(self.test_set)):
                 if self.valid_test(match=match):
                     self.verified_test_set.append(match)
+                    self.valid_matches += 1
 
         # Predict OpenSkill Matches
         title = HTML(f'<style fg="Blue">Predicting OpenSkill Matches:</style>')
@@ -129,22 +105,15 @@ class Rank:
             ):
                 self.predict_openskill(match=match)
 
-        # Predict TrueSkill Matches
-        title = HTML(f'<style fg="Blue">Predicting TrueSkill Matches:</style>')
-        with ProgressBar(title=title) as progress_bar:
-            for match in progress_bar(
-                self.verified_test_set, total=len(self.verified_test_set)
-            ):
[…]
