### 2023-08-12 ffd1f0b
Fixes philihp/openskill.js#453 and adds a better benchmark for ranks.
Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
