### 2023-08-12 06f264b
Add Large Dataset to Benchmarks (#106)
* Fixes philihp/openskill.js#453 and adds a better benchmark for ranks.

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Fix processor not found error

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Add changelog

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

---------

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
