+++ benchmark/processors/large.py
@@ -1,0 +1,359 @@
+import time
+from typing import Union
+
+import numpy as np
+import polars as pl
+import rbo
+import rich
+from prompt_toolkit.styles import Style
+from rich.progress import Progress
+from rich.table import Table
+from sklearn.model_selection import train_test_split
+
+from openskill.models import (
+    BradleyTerryFull,
+    BradleyTerryPart,
+    PlackettLuce,
+    ThurstoneMostellerFull,
+    ThurstoneMostellerPart,
+)
+
+
+class Large:
+    def __init__(
+        self,
+        path,
+        seed: int,
+        minimum_matches: int,
+        model: Union[
+            BradleyTerryFull,
+            BradleyTerryPart,
+            PlackettLuce,
+            ThurstoneMostellerFull,
+            ThurstoneMostellerPart,
+        ] = PlackettLuce,
+    ):
+        self.seed = seed
+        self.minimum_matches = minimum_matches
+        self.model = model
+
+        # Progress Bar Styles
+        self.style = Style.from_dict(
+            {
+                "label": "bg:#ffff00 #000000",
+                "percentage": "bg:#ffff00 #000000",
+                "current": "#448844",
+                "bar": "",
+            }
+        )
+
+        # Data
+        self.matches = {}
+        self.verified_matches = []
+
+        # Counters
+        self.match_count = {}
+        self.available_matches = 0
+        self.valid_matches = 0
+        self.openskill_correct_predictions = 0
+        self.openskill_incorrect_predictions = 0
+        self.rbo_scores = []
+
+        # Post Verification of Data
+        self.verified_matches = []
+        self.verified_test_set = []
+
+        # Split Data
+        self.training_set = []
+        self.test_set = []
+
+        # Ratings
+        self.openskill_players = {}
+
+        # Time Spent
+        self.openskill_time = None
+
+        # Generate Match Data
+        chunk_size = 10**5
+        total_size = 13849287
+        total_chunks = int(round(total_size / chunk_size))
+        reader = pl.read_csv_batched(path, batch_size=chunk_size)
+        batches = reader.next_batches(1)
+
+        with Progress() as progress:
+            loading_task = progress.add_task(
+                description="[red]Processing Chunks:", total=total_chunks
+            )
+            chunk_task = progress.add_task(
+                description="[red]Parsing and Transforming Data:", total=chunk_size
+            )
+            while batches:
+                for chunk in batches:
+                    progress.reset(chunk_task)
+                    progress.update(chunk_task, total=len(chunk))
+                    for raw_player in chunk.rows(named=True):
+                        player = {
+                            "name": raw_player["player_name"],
+                            "assists": raw_player["player_assists"],
+                            "kills": raw_player["player_kills"],
+                        }
+                        team = {
+                            "id": raw_player["team_id"],
+                            "match": raw_player["match_id"],
+                            "rank": raw_player["team_placement"],
+                            "size": raw_player["party_size"],
+                            "players": {},
+                        }
+                        match = {
+                            "id": raw_player["match_id"],
+                            "size": raw_player["game_size"],
+                            "teams": {},
+                        }
+                        match_id = match["id"]
+                        team_id = team["id"]
+                        player_name = player["name"]
+                        if match_id in self.matches:
+                            if team in self.matches[match_id]["teams"].values():
+                                self.matches[match_id]["teams"][team_id]["players"][
+                                    player_name
+                                ] = player
+                            else:
+                                self.matches[match_id]["teams"][team_id] = team
+                                self.matches[match_id]["teams"][team_id]["players"][
+                                    player_name
+                                ] = player
+                        else:
+                            self.matches[match_id] = match
[…]
