+++ changes/106.bugfix.rst
@@ -1,0 +1,1 @@
+Fixes philihp/openskill.js#453
