### 2023-08-12 a7b9058
Update CHANGELOG.rst
Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
