+++ changes/105.doc.rst
@@ -1,1 +1,0 @@
-Integrates Algolia Docsearch to Sphinx Docs
