+++ changes/106.bugfix.rst
@@ -1,1 +1,0 @@
-Fixes philihp/openskill.js#453
