+++ pyproject.toml
@@ -79,8 +79,10 @@ Tracker = "https://github.com/OpenDebates/openskill.py/issues"
 [tool.pdm.dev-dependencies]
 docs = [
     "sphinx~=7.0",
-    "shibuya~=2023.7",
+    "nbsphinx~=0.9",
     "pygments~=2.15",
+    "shibuya~=2023.7",
+    "ipykernel~=6.25",
     "myst-parser~=2.0",
     "sphinx-intl~=2.1",
     "sphinx-favicon~=1.0",
