+++ pyproject.toml
@@ -103,7 +103,7 @@ release = [
     "twine~=4.0",
     "build~=0.10",
     "isort~=5.12",
-    "black~=23.7",
+    "black[jupyter]~=23.7",
     "codecov~=2.1",
     "towncrier~=23.6",
     "bump2version~=1.0",
