### 2023-08-13 ef3011c
SEO Optimizations (#107)
* Improve SEO and add FAQ to docs.

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Format with black

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Format ipynb files with Black

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

---------

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
