+++ docs/source/api/openskill/models/weng_lin/thurstone_mosteller_part/index.rst
@@ -91,7 +91,7 @@ Classes
          A_q = |\{s: r(s) = r(q)\}|, q = 1,...,k
 
       :param team_ratings: The whole rating of a list of teams in a game.
-      :return: A list of integers.
+      :return: A list of Decimals.
 
 
    .. py:method:: _c(team_ratings)
@@ -162,7 +162,7 @@ Classes
 
       :param c: The square root of the collective team sigma.
 
-      :return: A list of integers.
+      :return: A list of Decimals.
 
 
    .. py:method:: create_rating(rating, name = None)
@@ -219,10 +219,10 @@ Classes
       :param teams: A list of teams where each team is a list of
                     :class:`ThurstoneMostellerPartRating` objects.
 
-      :param ranks: A list of integers where the lower values
+      :param ranks: A list of Decimals where the lower values
                     represent winners.
 
-      :param scores: A list of integers where higher values
+      :param scores: A list of Decimals where higher values
                     represent winners.
 
       :param tau: Additive dynamics parameter that prevents sigma from
