+++ CONTRIBUTORS.md
@@ -1,4 +1,4 @@
-## Contributors
+# Contributors
 
 Thanks goes to these wonderful people ([emoji key](https://allcontributors.org/docs/en/emoji-key)):
 
