+++ .bumpversion.cfg
@@ -1,15 +1,15 @@
 [bumpversion]
-current_version = 5.0.1
+current_version = 5.0.2
 commit = False
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)(\-(?P<release>[a-z]+)\.(?P<build>\d+))?
-serialize = 
+serialize =
 	{major}.{minor}.{patch}-{release}.{build}
 	{major}.{minor}.{patch}
 
 [bumpversion:part:release]
 optional_value = gamma
-values = 
+values =
 	alpha
 	beta
 	gamma
