+++ pyproject.toml
@@ -42,7 +42,7 @@ strict = true
 
 [project]
 name = "openskill"
-version = "5.0.1"
+version = "5.0.2"
 authors = [
     {name = "Vivek Joshy", email = "vivek@opendebates.net"},
 ]
