### 2023-08-13 a03847b
Correct changelog version
Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
