+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ This file is updated every release since v1.0.0 with the use of towncrier from t
 
 .. towncrier release notes start
 
-Openskill 5.0.1 (2023-08-13)
+Openskill 5.0.2 (2023-08-13)
 ============================
 
 Documentation Improvements
