### 2023-08-21 30fa6e5
Speedup deepcopy
Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
