+++ changes/108.feature.rst
@@ -1,0 +1,1 @@
+Optimize rating objects for deepcopy
