+++ tox.ini
@@ -1,7 +1,7 @@
 [gh]
 python =
     3.12-dev: python3.12
-    3.11: lint, build, type, py3.11
+    3.11: py3.11, lint, build, type
     3.10: py3.10
     3.9: py3.9
     3.8: py3.8
