### 2023-08-24 a037bc2
Add mypy to tox
Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
