+++ tox.ini
@@ -15,6 +15,7 @@ requires =
 env_list =
     lint
     build
+    type
     py{3.8,3.9,3.10,3.11}
     py{3.8,3.9,3.10,3.11}-win32
     python3.12
