+++ .github/workflows/main.yml
@@ -16,7 +16,7 @@ jobs:
     name: ${{ matrix.os }} / ${{ matrix.python-version }} (${{ matrix.architecture }})
     runs-on: ${{ matrix.os }}-latest
     strategy:
-      fail-fast: true
+      fail-fast: false
       matrix:
         os: [Ubuntu, MacOS, Windows]
         python-version: ["3.8", "3.9", "3.10", "3.11"]
