+++ tox.ini
@@ -1,4 +1,4 @@
-[gh]
+[gh-actions]
 python =
     3.12-dev: python3.12
     3.11: py3.11, lint, build, type
