### 2023-08-24 a4439f7
Fixes #109 (#110)
* Fixes #109

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Optimize Imports

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Add changelog

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Perform isort

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Add mypy to tox

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Prevent Fail Fast

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Correct GitHub Actions for T
[…]
