+++ changes/108.feature.rst
@@ -1,1 +1,0 @@
-Optimize rating objects for deepcopy
