+++ changes/109.bugfix.rst
@@ -1,1 +1,0 @@
-model.rating() methods no longer revert to default model parameters when arguments are 0
