+++ .bumpversion.toml
@@ -1,0 +1,31 @@
+[tool.bumpversion]
+allow_dirty = true
+commit = false
+message = "Bump version: {current_version} → {new_version}"
+commit_args = ""
+tag = false
+sign_tags = false
+tag_name = "v{new_version}"
+tag_message = "Bump version: {current_version} → {new_version}"
+current_version = "5.1.0"
+parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)(\\-(?P<release>[a-z]+)\\.(?P<build>\\d+))?"
+serialize = [
+    "{major}.{minor}.{patch}-{release}.{build}",
+    "{major}.{minor}.{patch}"
+]
+search = "{current_version}"
+replace = "{new_version}"
+
+[tool.bumpversion.parts.release]
+values = [
+    "alpha",
+    "beta",
+    "gamma"
+]
+optional_value = "gamma"
+
+[[tool.bumpversion.files]]
+filename = "openskill/__init__.py"
+
+[[tool.bumpversion.files]]
+filename = "pyproject.toml"
