+++ .bumpversion.cfg
@@ -1,19 +1,0 @@
-[bumpversion]
-current_version = 5.0.2
-commit = False
-tag = False
-parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)(\-(?P<release>[a-z]+)\.(?P<build>\d+))?
-serialize =
-	{major}.{minor}.{patch}-{release}.{build}
-	{major}.{minor}.{patch}
-
-[bumpversion:part:release]
-optional_value = gamma
-values =
-	alpha
-	beta
-	gamma
-
-[bumpversion:part:build]
-
-[bumpversion:file:./openskill/__init__.py]
