### 2023-08-25 11a02e1
Update sphinxcontrib-bibtex requirement from ~=2.5.0 to ~=2.6.0
Updates the requirements on [sphinxcontrib-bibtex](https://github.com/mcmtroffaes/sphinxcontrib-bibtex) to permit the latest version.
- [Changelog](https://github.com/mcmtroffaes/sphinxcontrib-bibtex/blob/develop/CHANGELOG.rst)
- [Commits](https://github.com/mcmtroffaes/sphinxcontrib-bibtex/compare/2.5.0...2.6.0)

---
updated-dependencies:
- dependency-name: sphinxcontrib-bibtex
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
