+++ docs/doc_requires.txt
@@ -9,5 +9,5 @@ sphinx_favicon~=1.0
 sphinx-copybutton~=0.5
 sphinx-autoapi~=2.1.1
 sphinxext-opengraph~=0.8
-sphinxcontrib-bibtex~=2.5.0
+sphinxcontrib-bibtex~=2.6.1
 sphinx-autodoc-typehints~=1.24.0
