### 2023-09-27 7b9adbf
Update sphinx-autoapi requirement from ~=2.1.1 to ~=3.0.0
Updates the requirements on [sphinx-autoapi](https://github.com/readthedocs/sphinx-autoapi) to permit the latest version.
- [Changelog](https://github.com/readthedocs/sphinx-autoapi/blob/main/CHANGELOG.rst)
- [Commits](https://github.com/readthedocs/sphinx-autoapi/compare/v2.1.1...v3.0.0)

---
updated-dependencies:
- dependency-name: sphinx-autoapi
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
