+++ docs/doc_requires.txt
@@ -7,7 +7,7 @@ myst_parser~=2.0  # Must Be Underscore, not Hyphen
 sphinx-intl~=2.1
 sphinx_favicon~=1.0
 sphinx-copybutton~=0.5
-sphinx-autoapi~=2.1.1
+sphinx-autoapi~=3.0.0
 sphinxext-opengraph~=0.8
 sphinxcontrib-bibtex~=2.5.0
 sphinx-autodoc-typehints~=1.24.0
