### 2023-09-29 #5901 @editorialbot
[REVIEW]: OpenSkill: A faster asymmetric multi-team, multiplayer rating system
**Submitting author:** @vivekjoshy (<a href="http://orcid.org/0000-0003-2443-8827">Vivek Joshy</a>)
**Repository:** https://github.com/OpenDebates/openskill.py
**Branch with paper.md** (empty if default branch): main
**Version:**  v5.1.0
