### 2023-10-15 #5901 @Naeemkh
Hello @vivekjoshy, please add the Open Journals GitHub Action to the package repository, either on the main branch or any other branch? This will help us see if the paper is generated successfully. For your reference, here are the guidelines: [JOSS Documentation](https://joss.readthedocs.io/en/latest/submitting.html?highlight=github%20action#github-action). Thank you.
