### 2023-10-15 #5901 @vivekjoshy
@Naeemkh it's already set to generate on each commit as seen here:

https://github.com/OpenDebates/openskill.py/actions/runs/5966197114
