### 2023-10-19 #5901 @xuanxu
When editorialbot is trying to clone the repository this error message is generated:
```
Error downloading benchmark/data/chess.csv
Error downloading benchmark/data/overwatch.jsonl
Error downloading benchmark/data/pubg.7z

batch response: This repository is over its data quota.
Account responsible for LFS bandwidth should purchase more data packs to restore access.
```
