### 2023-10-19 #5901 @vivekjoshy
@xuanxu There should be pointer files for the benchmarks if you use Git LFS. Unfortunately, the repo gets cloned quite a lot and there would be no point in topping up the data quota. I could always just delete the benchmark files or move it out to Kaggle since they're so large. That would also solve the issue.
