### 2023-10-21 OpenDebates/openskill.py#115 @Naeemkh
Software Paper Review: Suggestions for Clarity and Completeness
Please consider the following in drafting the software manuscript:
- Provide more contextual information about the problem you are solving in the paper and software, targeting software engineers and researchers who may not have specialized knowledge in the domain.
- There is a typo in the summary: know --> known.
- For claims about the software being "Faster" and "accurate", please offer supporting evidence like benchmarks, examples, or descriptions of the steps taken to achieve these qualities.
- Some of the limitations of the software are mentioned in the FAQ section of the documentation. It would be beneficial to dedicate a discussion section in the paper that covers what the package can and cannot do, as well as future development plans.
- The paper should give appropriate credit to the OpenSkills.js package.
- Please incorporate at least one example in the paper that demonstrates how to use the package, enabling readers to start using it quickly.

This issue is related to this submission: https://github.com/openjournals/joss-reviews/issues/5901
