### 2023-10-21 #5901 @Naeemkh
> @xuanxu There should be pointer files for the benchmarks if you use Git LFS. Unfortunately, the repo gets cloned quite a lot and there would be no point in topping up the data quota. I could always just delete the benchmark files or move it out to Kaggle since they're so large. That would also solve the issue.

@vivekjoshy, Please let me know where I can download the benchmark data.
