### 2023-10-22 #5901 @vivekjoshy
@Naeemkh I've uploaded the data files for Draw and Win benchmarks here:

https://www.kaggle.com/datasets/daegontaven/overwatch-dataset
https://www.kaggle.com/datasets/daegontaven/chess-dataset

It is currently not documented that the benchmark for Large, has an error in it. I've uploaded a corrected version of the benchmarks for the large dataset (pubg) here instead:

https://www.kaggle.com/datasets/daegontaven/processed-pubg

There is also an online notebook that can be viewed for it that uses the corrected dataset (please see version 25 once it's done running):

https://www.kaggle.com/code/daegontaven/openskill-benchmarks
