### 2023-10-22 ef0dec4
Update paper to include suggestions by reviewer
Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
