+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 # Summary
 
 Online ranking has been a long-standing issue due to many players being stuck with a
-particular rank, know colloquially as "Elo Hell" [@Elo]. Another major issue
+particular rank, known colloquially as "Elo Hell" [@Elo]. Another major issue
 is that most ranking systems are designed for 1v1 games, and do not scale well
 to multiplayer games. If they are designed for that purpose, the algorithms that
 are available are inefficient. This library is designed to  address all of these
@@ -31,10 +31,21 @@ problems while being faster than and just as accurate as proprietary algorithms.
 # Statement of need
 
 Bayesian inference of skill ratings from game outcomes is a crucial aspect of online
-video game development and research. The OpenSkill library provides a comprehensive
-set of models and algorithms to facilitate this process. It is primarily meant to be used
-by video game developers and researchers working on multi-agent scripting environments
-such as Neural MMO [@suarez2019neural]. Derived from the research paper by Weng and Lin [@JMLR:v12:weng11a],
+video game development and research. This is usually challenging because the players' performance changes
+over time and also varies based on who they are competing against. Our project primarily targets game developers
+and researchers interested in ranking players fairly and accurately. Nevertheless, the problem that the
+software solves is applicable to any context where you have multiple players or entities and you need to track
+their skills over time while they compete against each other.
+
+The OpenSkill library furnishes a versatile suite of models and algorithms designed to support a broad spectrum
+of applications. While popular use cases include assisting video game developers and researchers dealing with
+multi-agent scripting environments like Neural MMO [@suarez2019neural], its practical use extends far beyond this
+particular domain. For instance, it finds substantial utilization in recommendation systems, where it efficiently
+gauges unique user behaviors and preferences to suggest personalized recommendations. The matchmaking mechanisms
+in ranking of sports players as seen by Opta Analyst [@Rico_2022] and dating apps are another area where OpenSkill
+proves crucial, ensuring an optimal pairing based on the comparative ranking of user profiles' competencies.
+
+Derived from the research paper by Weng and Lin [@JMLR:v12:weng11a],
 OpenSkill offers a pure Python implementation of their models. Similar to TrueSkill,
 this framework is specifically designed for asymmetric multi-faction multiplayer games.
 However, OpenSkill boasts several advantages over proprietary models like TrueSkill [@herbrich2006trueskill].
@@ -50,8 +61,68 @@ sacrificing a certain level of accuracy. On the other hand, full pairing models 
 all available information within the paired data to make precise rating updates at the cost
 of increased computational requirements.
 
+# Usage
+
+To install the library simply `pip install openskill` and import the library. An conventional example of usage is given below:
+
+```python
+
+>>> from openskill.models import PlackettLuce
+>>> model = PlackettLuce()
+>>> model.rating()
+PlackettLuceRating(mu=25.0, sigma=8.333333333333334)
+>>> r = model.rating
+>>> [[a, b], [x, y]] = [[r(), r()], [r(), r()]]
+>>> [[a, b], [x, y]] = model.rate([[a, b], [x, y]])
+>>> a
+PlackettLuceRating(mu=26.964294621803063, sigma=8.177962604389991)
+>>> x
+PlackettLuceRating(mu=23.035705378196937, sigma=8.177962604389991)
+>>> (a == b) and (x == y)
+True
+```
+
+Each player has a `mu` and a `sigma` value corresponding to their skill ($\mu$) and uncertainty ($\sigma$) in skill. Comparisons between two players can be done by calling the `ordinal()` method.
+In this case it would be on the instances of `PlackettLuceRating`.
+
+# Benchmarks
+
[…]
