+++ paper/paper.bib
@@ -1,12 +1,12 @@
 @article{JMLR:v12:weng11a,
-  author  = {Ruby C. Weng and Chih-Jen Lin},
-  title   = {A Bayesian Approximation Method for Online Ranking},
-  journal = {Journal of Machine Learning Research},
-  year    = {2011},
-  volume  = {12},
-  number  = {9},
-  pages   = {267--300},
-  url     = {http://jmlr.org/papers/v12/weng11a.html}
+author  = {Ruby C. Weng and Chih-Jen Lin},
+title   = {A Bayesian Approximation Method for Online Ranking},
+journal = {Journal of Machine Learning Research},
+year    = {2011},
+volume  = {12},
+number  = {9},
+pages   = {267--300},
+url     = {http://jmlr.org/papers/v12/weng11a.html}
 }
 @article{Elo,
 title = {The psychology of esports players’ ELO Hell: Motivated bias in League of Legends and its impact on players’ overestimation of skill},
@@ -31,10 +31,45 @@ url = {https://www.microsoft.com/en-us/research/publication/trueskilltm-a-bayesi
 number = {MSR-TR-2006-80},
 }
 @misc{suarez2019neural,
-      title={Neural MMO: A Massively Multiagent Game Environment for Training and Evaluating Intelligent Agents},
-      author={Joseph Suarez and Yilun Du and Phillip Isola and Igor Mordatch},
-      year={2019},
-      eprint={1903.00784},
-      archivePrefix={arXiv},
-      primaryClass={cs.MA}
+title={Neural MMO: A Massively Multiagent Game Environment for Training and Evaluating Intelligent Agents},
+author={Joseph Suarez and Yilun Du and Phillip Isola and Igor Mordatch},
+year={2019},
+eprint={1903.00784},
+archivePrefix={arXiv},
+primaryClass={cs.MA}
+}
+@misc{Rico_2022,
+title={Her ranking is a 10 but her skill rating says...},
+url={https://theanalyst.com/eu/2022/08/true-tennis-skill-ratings/},
+journal={The Analyst},
+publisher={Opta Analyst},
+author={Rico, Ysabel González},
+year={2022},
+month={Aug}
+}
+@article{10.1145/1852102.1852106,
+author = {Webber, William and Moffat, Alistair and Zobel, Justin},
+title = {A Similarity Measure for Indefinite Rankings},
+year = {2010},
+issue_date = {November 2010},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+volume = {28},
+number = {4},
+issn = {1046-8188},
+url = {https://doi.org/10.1145/1852102.1852106},
+doi = {10.1145/1852102.1852106},
[…]
