### 2023-10-23 135058a
fix typo (#117)
Signed-off-by: Jack McIvor <jacktmcivor@gmail.com>
