### 2023-10-23 52bd90e
Remove unneccessary double unions
