+++ openskill/models/weng_lin/plackett_luce.py
@@ -377,7 +377,7 @@ class PlackettLuce:
 
     @staticmethod
     def create_rating(
-        rating: List[Union[Union[int, float]]], name: Optional[str] = None
+        rating: List[Union[int, float]], name: Optional[str] = None
     ) -> PlackettLuceRating:
         """
         Create a :class:`PlackettLuceRating` object from a list of `mu`
