+++ openskill/models/weng_lin/bradley_terry_part.py
@@ -382,7 +382,7 @@ class BradleyTerryPart:
 
     @staticmethod
     def create_rating(
-        rating: List[Union[Union[int, float]]], name: Optional[str] = None
+        rating: List[Union[int, float]], name: Optional[str] = None
     ) -> BradleyTerryPartRating:
         """
         Create a :class:`BradleyTerryPartRating` object from a list of `mu`
