### 2023-10-23 1a451b0
Replace typing.Union[int, float] -> float
