+++ openskill/models/weng_lin/thurstone_mosteller_full.py
@@ -33,8 +33,8 @@ class ThurstoneMostellerFullRating:
 
     def __init__(
         self,
-        mu: Union[int, float],
-        sigma: Union[int, float],
+        mu: float,
+        sigma: float,
         name: Optional[str] = None,
     ):
         r"""
@@ -57,8 +57,8 @@ class ThurstoneMostellerFullRating:
         self.id: str = uuid.uuid4().hex.lower()
         self.name: Optional[str] = name
 
-        self.mu: Union[int, float] = mu
-        self.sigma: Union[int, float] = sigma
+        self.mu: float = mu
+        self.sigma: float = sigma
 
     def __repr__(self) -> str:
         return f"ThurstoneMostellerFullRating(mu={self.mu}, sigma={self.sigma})"
@@ -143,7 +143,7 @@ class ThurstoneMostellerFullRating:
                 "You can only compare ThurstoneMostellerFullRating objects with each other."
             )
 
-    def ordinal(self, z: Union[int, float] = 3.0) -> Union[int, float]:
+    def ordinal(self, z: float = 3.0) -> float:
         r"""
         A single scalar value that represents the player's skill where their
         true skill is 99.7% likely to be higher.
@@ -163,8 +163,8 @@ class ThurstoneMostellerFullTeamRating:
 
     def __init__(
         self,
-        mu: Union[int, float],
-        sigma_squared: Union[int, float],
+        mu: float,
+        sigma_squared: float,
         team: Sequence[ThurstoneMostellerFullRating],
         rank: int,
     ):
@@ -217,13 +217,13 @@ class ThurstoneMostellerFullTeamRating:
 
 
 def _gamma(
-    c: Union[int, float],
+    c: float,
     k: int,
-    mu: Union[int, float],
-    sigma_squared: Union[int, float],
+    mu: float,
+    sigma_squared: float,
     team: Sequence[ThurstoneMostellerFullRating],
     rank: int,
-) -> Union[int, float]:
+) -> float:
     """
     Default gamma function for Thurstone-Mosteller Full Pairing.
 
@@ -259,22 +259,22 @@ class ThurstoneMostellerFull:
 
     def __init__(
         self,
-        mu: Union[int, float] = 25.0,
-        sigma: Union[int, float] = 25.0 / 3.0,
-        beta: Union[int, float] = 25.0 / 6.0,
-        kappa: Union[int, float] = 0.0001,
+        mu: float = 25.0,
+        sigma: float = 25.0 / 3.0,
+        beta: float = 25.0 / 6.0,
+        kappa: float = 0.0001,
         gamma: Callable[
             [
-                Union[int, float],
+                float,
                 int,
-                Union[int, float],
-                Union[int, float],
+                float,
+                float,
                 Sequence[ThurstoneMostellerFullRating],
                 int,
             ],
-            Union[int, float],
+            float,
         ] = _gamma,
-        tau: Union[int, float] = 25.0 / 300.0,
+        tau: float = 25.0 / 300.0,
         limit_sigma: bool = False,
     ):
         r"""
@@ -319,23 +319,23 @@ class ThurstoneMostellerFull:
 
         """
         # Model Parameters
-        self.mu: Union[int, float] = float(mu)
-        self.sigma: Union[int, float] = float(sigma)
-        self.beta: Union[int, float] = beta
-        self.kappa: Union[int, float] = float(kappa)
+        self.mu: float = float(mu)
+        self.sigma: float = float(sigma)
+        self.beta: float = beta
+        self.kappa: float = float(kappa)
         self.gamma: Callable[
             [
-                Union[int, float],
+                float,
                 int,
-                Union[int, float],
[…]
