+++ openskill/models/weng_lin/plackett_luce.py
@@ -613,9 +613,7 @@ class PlackettLuce:
         return math.sqrt(collective_team_sigma)
 
     @staticmethod
-    def _sum_q(
-        team_ratings: List[PlackettLuceTeamRating], c: float
-    ) -> List[float]:
+    def _sum_q(team_ratings: List[PlackettLuceTeamRating], c: float) -> List[float]:
         r"""
         Sum up all the values of :code:`mu / c` raised to :math:`e`.
 
@@ -724,9 +722,7 @@ class PlackettLuce:
             result.append(intermediate_result_per_team)
         return result
 
-    def predict_win(
-        self, teams: List[List[PlackettLuceRating]]
-    ) -> List[float]:
+    def predict_win(self, teams: List[List[PlackettLuceRating]]) -> List[float]:
         r"""
         Predict how likely a match up against teams of one or more players
         will go. This algorithm has a time complexity of
