+++ openskill/models/weng_lin/bradley_terry_part.py
@@ -619,9 +619,7 @@ class BradleyTerryPart:
         return math.sqrt(collective_team_sigma)
 
     @staticmethod
-    def _sum_q(
-        team_ratings: List[BradleyTerryPartTeamRating], c: float
-    ) -> List[float]:
+    def _sum_q(team_ratings: List[BradleyTerryPartTeamRating], c: float) -> List[float]:
         r"""
         Sum up all the values of :code:`mu / c` raised to :math:`e`.
 
@@ -740,9 +738,7 @@ class BradleyTerryPart:
 
         return list(map(lambda i: i_map(i[0], i[1]), zip(team_ratings, adjacent_teams)))
 
-    def predict_win(
-        self, teams: List[List[BradleyTerryPartRating]]
-    ) -> List[float]:
+    def predict_win(self, teams: List[List[BradleyTerryPartRating]]) -> List[float]:
         r"""
         Predict how likely a match up against teams of one or more players
         will go. This algorithm has a time complexity of
@@ -798,9 +794,7 @@ class BradleyTerryPart:
             )
         ]
 
-    def predict_draw(
-        self, teams: List[List[BradleyTerryPartRating]]
-    ) -> float:
+    def predict_draw(self, teams: List[List[BradleyTerryPartRating]]) -> float:
         r"""
         Predict how likely a match up against teams of one or more players
         will draw. This algorithm has a time complexity of
