### 2023-10-23 1c534ce
Fix/minor typings (#118)
* Remove unneccessary double unions

* Replace typing.Union[int, float] -> float

* Remove unused imports

* Format

---------

Co-authored-by: Jack McIvor <j.mcivor@draftkings.com>
