+++ openskill/models/weng_lin/bradley_terry_full.py
@@ -8,7 +8,7 @@ import itertools
 import math
 import uuid
 from functools import reduce
-from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Type, Union
+from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Type
 
 from openskill.models.common import _rank_data, _unary_minus
 from openskill.models.weng_lin.common import _unwind, phi_major, phi_major_inverse
@@ -25,8 +25,8 @@ class BradleyTerryFullRating:
 
     def __init__(
         self,
-        mu: Union[int, float],
-        sigma: Union[int, float],
+        mu: float,
+        sigma: float,
         name: Optional[str] = None,
     ):
         r"""
@@ -49,8 +49,8 @@ class BradleyTerryFullRating:
         self.id: str = uuid.uuid4().hex.lower()
         self.name: Optional[str] = name
 
-        self.mu: Union[int, float] = mu
-        self.sigma: Union[int, float] = sigma
+        self.mu: float = mu
+        self.sigma: float = sigma
 
     def __repr__(self) -> str:
         return f"BradleyTerryFullRating(mu={self.mu}, sigma={self.sigma})"
@@ -133,7 +133,7 @@ class BradleyTerryFullRating:
                 "You can only compare BradleyTerryFullRating objects with each other."
             )
 
-    def ordinal(self, z: Union[int, float] = 3.0) -> Union[int, float]:
+    def ordinal(self, z: float = 3.0) -> float:
         r"""
         A single scalar value that represents the player's skill where their
         true skill is 99.7% likely to be higher.
@@ -153,8 +153,8 @@ class BradleyTerryFullTeamRating:
 
     def __init__(
         self,
-        mu: Union[int, float],
-        sigma_squared: Union[int, float],
+        mu: float,
+        sigma_squared: float,
         team: Sequence[BradleyTerryFullRating],
         rank: int,
     ):
@@ -207,13 +207,13 @@ class BradleyTerryFullTeamRating:
 
 
 def _gamma(
-    c: Union[int, float],
+    c: float,
     k: int,
-    mu: Union[int, float],
-    sigma_squared: Union[int, float],
+    mu: float,
+    sigma_squared: float,
     team: Sequence[BradleyTerryFullRating],
     rank: int,
-) -> Union[int, float]:
+) -> float:
     """
     Default gamma function for Bradley-Terry Full Pairing.
 
@@ -248,22 +248,22 @@ class BradleyTerryFull:
 
     def __init__(
         self,
-        mu: Union[int, float] = 25.0,
-        sigma: Union[int, float] = 25.0 / 3.0,
-        beta: Union[int, float] = 25.0 / 6.0,
-        kappa: Union[int, float] = 0.0001,
+        mu: float = 25.0,
+        sigma: float = 25.0 / 3.0,
+        beta: float = 25.0 / 6.0,
+        kappa: float = 0.0001,
         gamma: Callable[
             [
-                Union[int, float],
+                float,
                 int,
-                Union[int, float],
-                Union[int, float],
+                float,
+                float,
                 Sequence[BradleyTerryFullRating],
                 int,
             ],
-            Union[int, float],
+            float,
         ] = _gamma,
-        tau: Union[int, float] = 25.0 / 300.0,
+        tau: float = 25.0 / 300.0,
         limit_sigma: bool = False,
     ):
         r"""
@@ -308,23 +308,23 @@ class BradleyTerryFull:
 
         """
         # Model Parameters
-        self.mu: Union[int, float] = float(mu)
-        self.sigma: Union[int, float] = float(sigma)
-        self.beta: Union[int, float] = beta
-        self.kappa: Union[int, float] = float(kappa)
[…]
