+++ openskill/models/weng_lin/bradley_terry_part.py
@@ -8,7 +8,7 @@ import itertools
 import math
 import uuid
 from functools import reduce
-from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Type, Union
+from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Type
 
 from openskill.models.common import _rank_data, _unary_minus
 from openskill.models.weng_lin.common import (
@@ -30,8 +30,8 @@ class BradleyTerryPartRating:
 
     def __init__(
         self,
-        mu: Union[int, float],
-        sigma: Union[int, float],
+        mu: float,
+        sigma: float,
         name: Optional[str] = None,
     ):
         r"""
@@ -54,8 +54,8 @@ class BradleyTerryPartRating:
         self.id: str = uuid.uuid4().hex.lower()
         self.name: Optional[str] = name
 
-        self.mu: Union[int, float] = mu
-        self.sigma: Union[int, float] = sigma
+        self.mu: float = mu
+        self.sigma: float = sigma
 
     def __repr__(self) -> str:
         return f"BradleyTerryPartRating(mu={self.mu}, sigma={self.sigma})"
@@ -138,7 +138,7 @@ class BradleyTerryPartRating:
                 "You can only compare BradleyTerryPartRating objects with each other."
             )
 
-    def ordinal(self, z: Union[int, float] = 3.0) -> Union[int, float]:
+    def ordinal(self, z: float = 3.0) -> float:
         r"""
         A single scalar value that represents the player's skill where their
         true skill is 99.7% likely to be higher.
@@ -158,8 +158,8 @@ class BradleyTerryPartTeamRating:
 
     def __init__(
         self,
-        mu: Union[int, float],
-        sigma_squared: Union[int, float],
+        mu: float,
+        sigma_squared: float,
         team: Sequence[BradleyTerryPartRating],
         rank: int,
     ):
@@ -212,13 +212,13 @@ class BradleyTerryPartTeamRating:
 
 
 def _gamma(
-    c: Union[int, float],
+    c: float,
     k: int,
-    mu: Union[int, float],
-    sigma_squared: Union[int, float],
+    mu: float,
+    sigma_squared: float,
     team: Sequence[BradleyTerryPartRating],
     rank: int,
-) -> Union[int, float]:
+) -> float:
     """
     Default gamma function for Bradley-Terry Partial Pairing.
 
@@ -253,22 +253,22 @@ class BradleyTerryPart:
 
     def __init__(
         self,
-        mu: Union[int, float] = 25.0,
-        sigma: Union[int, float] = 25.0 / 3.0,
-        beta: Union[int, float] = 25.0 / 6.0,
-        kappa: Union[int, float] = 0.0001,
+        mu: float = 25.0,
+        sigma: float = 25.0 / 3.0,
+        beta: float = 25.0 / 6.0,
+        kappa: float = 0.0001,
         gamma: Callable[
             [
-                Union[int, float],
+                float,
                 int,
-                Union[int, float],
-                Union[int, float],
+                float,
+                float,
                 Sequence[BradleyTerryPartRating],
                 int,
             ],
-            Union[int, float],
+            float,
         ] = _gamma,
-        tau: Union[int, float] = 25.0 / 300.0,
+        tau: float = 25.0 / 300.0,
         limit_sigma: bool = False,
     ):
         r"""
@@ -313,23 +313,23 @@ class BradleyTerryPart:
 
         """
         # Model Parameters
-        self.mu: Union[int, float] = float(mu)
-        self.sigma: Union[int, float] = float(sigma)
-        self.beta: Union[int, float] = beta
-        self.kappa: Union[int, float] = float(kappa)
+        self.mu: float = float(mu)
[…]
