+++ openskill/models/common.py
@@ -1,10 +1,10 @@
 """
 Common functions for all models.
 """
-from typing import Any, List, Union
+from typing import Any, List
 
 
-def _unary_minus(number: Union[int, float]) -> Union[int, float]:
+def _unary_minus(number: float) -> float:
     """
     Takes value of a number and makes it negative.
 

[…]
