### 2023-11-22 OpenDebates/openskill.py#120 @matt-graham
Community guidelines for reporting issues and support queries
Raising as part of JOSS review openjournals/joss-reviews/issues/5901

Ideally you should some clear and easily findable guidelines for how to report issues and seek support with the software.

This section in the user manual page in the documentation

https://github.com/OpenDebates/openskill.py/blob/bc96949febda5a285af34a901d7128c95501e9a4/docs/source/manual.rst?plain=1#L10-L13

already partially fits the bill, but

 1. It took a little while for me to find - I would put it somewhere more prominent, for example in a dedicated top-level documentation page or in README
 2. The reference to 'discussions section' is not very clear - I assume this means GitHub Discussions but from documentation website it wouldn't necessarily be clear how to get there so adding a link would be useful.
 3. A brief description and pointer to the GitHub issue tracker as (presumably) the correct place to report issues with the code, and perhaps some explanation of the different issue templates / categories.
