### 2023-11-22 OpenDebates/openskill.py#121 @matt-graham
Documenting how to access data for benchmarking
Raising as part of JOSS review openjournals/joss-reviews/issues/5901

As the data files are stored on Git LFS and the free LFS quota for this account seems to be regularly exceeded (see https://github.com/openjournals/joss-reviews/issues/5901#issuecomment-1770817582) it would be useful to document an alternative approach for accessing the data, ideally one which uses an open data repository which doesn't require subscribing to an account to download. While the datasets have been made available on Kaggle (https://github.com/openjournals/joss-reviews/issues/5901#issuecomment-1774019894) this is not currently documented in this repository and a Kaggle account is required to download. An open research data repository / archive like Zenodo would seem to be a better fit with [JOSS requirement](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) that the software should be _stored in a repository that can be cloned without registration_. While I don't think this strictly extends to data associated with the software, from a FAIR data and reproducibility perspective a service like Zenodo is much better than Kaggle.

A potentially even nicer approach would be to use a tool like [pooch](https://pypi.org/project/pooch/) to automate getting the data from a remote repository as part of running the benchmarks.
