### 2023-11-22 OpenDebates/openskill.py#122 @matt-graham
Clearer statement of need in documentation
Raising as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5901

The documentation / top-level README need a clearer statement of need / what problems the software is designed to solve and who the intended audience is.

The current summary at the begining of the README

> A faster and open license asymmetric multi-team, multiplayer rating system comparable to TrueSkill.

assumes knowledge of what a multiplayer rating system is and what TrueSkill is. There is also no specific mention of online gaming communities, which reading between the lines, seems to be the primary target audience.

The summary in the documentation index page

> This advanced rating system is faster, accurate, and flexible with an open license. It supports multiple teams, factions, and predicts outcomes. Elevate your gaming experience with OpenSkill - the superior alternative to TrueSkill.

similarly needs a bit more context and explanation.
