### 2023-11-29 #5901 @matt-graham
Just to update as I realised I didn't follow up here last week: my review checklist is now fully up to date in the sense that I've ticked off all items that I believe the current repository and paper satisfy, with the unticked items reflecting things I believe require some action before being satisfied, with details on the changes I think are needed raised in specific issues on the `openskill.py` repository. @vivekjoshy let me know if you need any clarification on the points I've raised.
