+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{JMLR:v12:weng11a,
 author  = {Ruby C. Weng and Chih-Jen Lin},
-title   = {A Bayesian Approximation Method for Online Ranking},
+title   = {A {B}ayesian Approximation Method for Online Ranking},
 journal = {Journal of Machine Learning Research},
 year    = {2011},
 volume  = {12},
@@ -23,10 +23,10 @@ abstract = {This paper examines the folk theory of ELO Hell, which stems from th
 }
 @techreport{herbrich2006trueskill,
 author = {Herbrich, Ralf and Graepel, Thore},
-title = {TrueSkill(TM): A Bayesian Skill Rating System},
+title = {TrueSkill(TM): A {B}ayesian Skill Rating System},
 year = {2006},
 month = {January},
-abstract = {We present a new Bayesian skill rating system which can be viewed as a generalisation of the Elo system used in Chess. The new system tracks the uncertainty about player skills, explicitly models draws, can deal with any number of competing entities and can infer individual skills from team results. Inference is performed by approximate message passing on a factor graph representation of the model. We present experimental evidence on the increased accuracy and convergence speed of the system compared to Elo and report on our experience with the new rating system running in a large-scale commercial online gaming service under the name of TrueSkill.},
+abstract = {We present a new {B}ayesian skill rating system which can be viewed as a generalisation of the Elo system used in Chess. The new system tracks the uncertainty about player skills, explicitly models draws, can deal with any number of competing entities and can infer individual skills from team results. Inference is performed by approximate message passing on a factor graph representation of the model. We present experimental evidence on the increased accuracy and convergence speed of the system compared to Elo and report on our experience with the new rating system running in a large-scale commercial online gaming service under the name of TrueSkill.},
 url = {https://www.microsoft.com/en-us/research/publication/trueskilltm-a-bayesian-skill-rating-system-2/},
 number = {MSR-TR-2006-80},
 }
@@ -72,4 +72,53 @@ url={https://github.com/philihp/openskill.js},
 journal={GitHub},
 author={Busby, Philihp},
 year={2023}, month={Apr}
+}
+@article{Glicko2,
+author = {Mark E. Glickman},
+title = {Dynamic paired comparison models with stochastic variances},
+journal = {Journal of Applied Statistics},
+volume = {28},
+number = {6},
+pages = {673-689},
+year = {2001},
+publisher = {Taylor & Francis},
+doi = {10.1080/02664760120059219},
+URL = {
+    https://doi.org/10.1080/02664760120059219
+},
+eprint = {
+    https://doi.org/10.1080/02664760120059219
+}
+}
+@book{elo_rating_1978,
+address = {New York},
+title = {The rating of chessplayers, past and present},
+isbn = {9780668047210},
+publisher = {Arco Pub},
+author = {Elo, Arpad E.},
+year = {1978},
+keywords = {Chess players, Rating of, Probabilities},
+}
+@misc{Lee_2018,
+title={An implementation of the TrueSkill rating system for Python},
+url={https://github.com/sublee/trueskill},
+journal={sublee/trueskill: An implementation of the TrueSkill rating system for Python},
+author={Lee, Heungsub},
+year={2018}, month={Sep}
+}
+@misc{Minka_Graepel_Herbrich_2013,
+title={Determining relative skills of players},
+author={Minka, Thomas and  Graepel, Thore K H and Herbrich, Ralf},
+year={2013},
+month={Sep}
+}
+@dataset{joshy_2023_overwatch,
+  author       = {Joshy, Vivek},
+  title        = {OverWatch Match Data},
+  month        = dec,
[…]
