+++ paper/paper.md
@@ -21,49 +21,104 @@ bibliography: paper.bib
 
 # Summary
 
-Online ranking has been a long-standing issue due to many players being stuck with a
-particular rank, known colloquially as "Elo Hell" [@Elo]. Another major issue
-is that most ranking systems are designed for 1v1 games, and do not scale well
-to multiplayer games. If they are designed for that purpose, the algorithms that
-are available are inefficient. This library is designed to  address all of these
-problems while being faster than and just as accurate as proprietary algorithms.
+In the realm of online gaming, player performance is often measured and
+compared through a system called online ranking. This system assigns a
+"rank" to players based on their performance and outcomes in games. A rank is
+a distinct level or position within a hierarchy that ostensibly represents a
+player's skill relative to others. One common criticism of this system is the
+phenomenon known as "Elo Hell", a situation where players find themselves
+trapped at a certain rank, unable to progress due to perceived flaws in how
+the ranking is calculated or due to the influence of team dynamics [@Elo].
+
+In the context of player ranking systems, several well-established models
+such as Elo and Glicko 2 have set the standard for performance measurement.
+However, a limitation becomes evident when these commonly used systems are
+considered for games that extend beyond two-player matchups.
+The Elo rating system [@elo_rating_1978], for one, was originally formulated
+for chess, a head-to-head game, and it struggles to adapt to the complex
+dynamics of multiplayer scenarios. Similarly, Glicko 2 [@Glicko2], while
+offering improvements in rating accuracy and accounting for player rating
+volatility, still falls short when tasked with accurately representing the
+individual contributions within a team-oriented game setting.
+
+This library represents a concrete implementation of an improved ranking
+system, specifically engineered to address the distinctive challenges of
+multiplayer gaming environments. It systematically corrects for the
+deficiencies of traditional ranking systems, providing a robust solution
+that ensures a fair and dynamic evaluation of each player’s skill level.
+In addition to delivering accuracy on par with implementations of
+proprietary models like TrueSkill [@herbrich2006trueskill], this system
+distinguishes itself through its enhanced speed in computing the ranks,
+facilitating a more responsive and gratifying player experience.
 
 # Statement of need
 
-Bayesian inference of skill ratings from game outcomes is a crucial aspect of online
-video game development and research. This is usually challenging because the players' performance changes
-over time and also varies based on who they are competing against. Our project primarily targets game developers
-and researchers interested in ranking players fairly and accurately. Nevertheless, the problem that the
-software solves is applicable to any context where you have multiple players or entities and you need to track
+Bayesian inference of skill ratings from game outcomes is a crucial aspect of
+online video game development and research. This is usually challenging
+because the players' performance changes over time and also varies based on
+who they are competing against. Our project primarily targets game developers
+and researchers interested in ranking players fairly and accurately.
+Nevertheless, the problem that the software solves is applicable to any
+context where you have multiple players or entities and you need to track
 their skills over time while they compete against each other.
 
-The OpenSkill library furnishes a versatile suite of models and algorithms designed to support a broad spectrum
-of applications. While popular use cases include assisting video game developers and researchers dealing with
-multi-agent scripting environments like Neural MMO [@suarez2019neural], its practical use extends far beyond this
-particular domain. For instance, it finds substantial utilization in recommendation systems, where it efficiently
-gauges unique user behaviors and preferences to suggest personalized recommendations. The matchmaking mechanisms
[…]
