+++ docs/source/index.rst
@@ -10,8 +10,9 @@
     :align: center
 
 
-This advanced rating system is faster, accurate, and flexible with an open license. It supports multiple teams, factions, and predicts outcomes. Elevate your gaming experience with OpenSkill - the superior alternative to TrueSkill.
+In the multifaceted world of online gaming, an accurate multiplayer rating system plays a crucial role. A multiplayer rating system measures and compares players' skill levels in competitive games to ensure balanced match-making, boosting overall gaming experiences. Currently, TrueSkill by Microsoft Research is a notable rating system, but gaming communities are yearning for faster, more adaptable alternatives.
 
+With the rise of varying team sizes in matches - known as asymmetric multi-team scenarios - a more sophisticated rating system is required. Recognizing this need, advancements towards developing an open-license, efficient rating system have been initiated. This upgraded system seeks to facilitate a broad range of game formats and team configurations, promising to elevate competitive play and matchmaking in online gaming to the next level.
 
 .. toctree::
    :maxdepth: 1

[…]
