### 2023-12-13 #5901 @xuanxu
The bot is still failing to clone the repository before doing the checkout to the `paper` branch because the Git LFS files are still present in the `main` branch.
```
Error downloading benchmark/data/chess.csv:
batch response: This repository is over its data quota.
Account responsible for LFS bandwidth should purchase more data packs to restore access.
```
