### 2023-12-13 #5901 @vissarion
@xuanxu are you reading this from the latest error?
https://github.com/openjournals/joss-papers/actions/runs/7192563120/job/19589242658
