### 2023-12-13 #5901 @xuanxu
@vissarion I just tried to `git clone` the repo locally and the command failed with that error message, I guess the bot got the same error.
