### 2023-12-13 #5901 @vissarion
Exactly, this is a different error.
