### 2023-12-13 #5901 @vissarion
Unless the download error is related to benchmark directory size and the CI error message is cryptic.
