### 2023-12-13 #5901 @danielskatz
@xuanxu - when I do this locally, it seems to work
```
% git clone  https://github.com/OpenDebates/openskill.py
Cloning into 'openskill.py'...
remote: Enumerating objects: 3353, done.
remote: Total 3353 (delta 459), reused 590 (delta 386), pack-reused 2652
Resolving deltas: 100% (2197/2197), done.
```
What am I missing that's failing?
