### 2023-12-19 #5901 @arfon
```
Cloning into 'openskill.py'...
remote: Enumerating objects: 3353, done.
remote: Total 3353 (delta 459), reused 590 (delta 386), pack-reused 2652
Resolving deltas: 100% (2197/2197), done.
git-lfs filter-process: git-lfs: command not found
fatal: the remote end hung up unexpectedly
warning: Clone succeeded, but checkout failed.
You can inspect what was checked out with 'git status'
and retry with 'git restore --source=HEAD :/'
```

I think this repository might use Git LFS under the hood and for some reason that's breaking the clone process? @vivekjoshy – any ideas what might be going on with your repository?
