### 2023-12-19 #5901 @vivekjoshy
@vivekjoshy yes it's using git-lfs in the main branch. But it's removed in the unmerged paper branch. Could be that the main branch is being cloned and checkout fails there maybe? I could merge the paper branch once the reviewers get back on the latest changes. Perhaps that will fix the issue. Another option could be I could just remove git-lfs from the main branch and see if that fixes it.
