### 2024-01-01 64ea87a
Update shibuya requirement from ~=2023.3 to ~=2024.1
Updates the requirements on [shibuya](https://github.com/lepture/shibuya) to permit the latest version.
- [Release notes](https://github.com/lepture/shibuya/releases)
- [Changelog](https://github.com/lepture/shibuya/blob/master/docs/changelog.rst)
- [Commits](https://github.com/lepture/shibuya/compare/2023.3.1...2024.1.1)

---
updated-dependencies:
- dependency-name: shibuya
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
