+++ docs/doc_requires.txt
@@ -1,7 +1,7 @@
 sphinx~=7.0
 nbsphinx~=0.9
 pygments~=2.15
-shibuya~=2023.3
+shibuya~=2024.1
 ipykernel~=6.25
 myst_parser~=2.0  # Must Be Underscore, not Hyphen
 sphinx-intl~=2.1
