### 2024-01-04 326f798
Incroporate latest suggestion given by Naeemkh.
https://github.com/openjournals/joss-reviews/issues/5901#issuecomment-1877180920

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
