+++ paper/paper.bib
@@ -106,12 +106,6 @@ journal={sublee/trueskill: An implementation of the TrueSkill rating system for 
 author={Lee, Heungsub},
 year={2018}, month={Sep}
 }
-@misc{Minka_Graepel_Herbrich_2013,
-title={Determining relative skills of players},
-author={Minka, Thomas and  Graepel, Thore K H and Herbrich, Ralf},
-year={2013},
-month={Sep}
-}
 @dataset{joshy_2023_overwatch,
   author       = {Joshy, Vivek},
   title        = {OverWatch Match Data},
@@ -121,4 +115,55 @@ month={Sep}
   version      = {1.0.0},
   doi          = {10.5281/zenodo.10359600},
   url          = {https://doi.org/10.5281/zenodo.10359600}
+}
+@inproceedings{Plackett_Luce,
+    author = {Guiver, John and Snelson, Edward},
+    title = {Bayesian Inference for Plackett-Luce Ranking Models},
+    year = {2009},
+    isbn = {9781605585161},
+    publisher = {Association for Computing Machinery},
+    address = {New York, NY, USA},
+    url = {https://doi.org/10.1145/1553374.1553423},
+    doi = {10.1145/1553374.1553423},
+    abstract = {This paper gives an efficient Bayesian method for inferring the parameters of a Plackett-Luce ranking model. Such models are parameterised distributions over rankings of a finite set of objects, and have typically been studied and applied within the psychometric, sociometric and econometric literature. The inference scheme is an application of Power EP (expectation propagation). The scheme is robust and can be readily applied to large scale data sets. The inference algorithm extends to variations of the basic Plackett-Luce model, including partial rankings. We show a number of advantages of the EP approach over the traditional maximum likelihood method. We apply the method to aggregate rankings of NASCAR racing drivers over the 2002 season, and also to rankings of movie genres.},
+    booktitle = {Proceedings of the 26th Annual International Conference on Machine Learning},
+    pages = {377–384},
+    numpages = {8},
+    location = {Montreal, Quebec, Canada},
+    series = {ICML '09}
+}
+@article{Bradley_Terry,
+    author = {BRADLEY, RALPH ALLAN and TERRY, MILTON E.},
+    title = "{RANK ANALYSIS OF INCOMPLETE BLOCK DESIGNS: THE METHOD OF PAIRED COMPARISONS}",
+    journal = {Biometrika},
+    volume = {39},
+    number = {3-4},
+    pages = {324-345},
+    year = {1952},
+    month = {12},
+    issn = {0006-3444},
+    doi = {10.1093/biomet/39.3-4.324},
+    url = {https://doi.org/10.1093/biomet/39.3-4.324},
+    eprint = {https://academic.oup.com/biomet/article-pdf/39/3-4/324/930466/39-3-4-324.pdf},
+}
+@misc{PhilihpOpenSkillElixir,
+	author = {Philihp Busby},
+	title = {{G}it{H}ub - philihp/openskill.ex: {E}lixir implementation of {W}eng-{L}in {B}ayesian ranking, a better, license-free alternative to {T}rue{S}kill --- github.com},
+	howpublished = {\url{https://github.com/philihp/openskill.ex}},
+	year = {2020},
+	note = {[Accessed 04-01-2024]},
+}
+@misc{BrezinaOpenSkillKotlin,
+	author = {Jan Brezina},
+	title = {{G}it{H}ub - brezinajn/openskill.kt --- github.com},
+	howpublished = {\url{https://github.com/brezinajn/openskill.kt}},
+	year = {2022},
+	note = {[Accessed 04-01-2024]},
+}
+@misc{Bstummeropenskilllua,
[…]
