### 2024-01-04 52ee70d
Fix capitalization in citation
Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
