+++ paper/paper.bib
@@ -133,8 +133,8 @@ year={2018}, month={Sep}
     series = {ICML '09}
 }
 @article{Bradley_Terry,
-    author = {BRADLEY, RALPH ALLAN and TERRY, MILTON E.},
-    title = "{RANK ANALYSIS OF INCOMPLETE BLOCK DESIGNS: THE METHOD OF PAIRED COMPARISONS}",
+    author = {Bradley, Ralph Allan and Terry, Milton E.},
+    title = "{Rank Analysis of incomplete block designs: {T}he method of paired comparisons}",
     journal = {Biometrika},
     volume = {39},
     number = {3-4},
