### 2024-01-04 #5901 @vivekjoshy
@Naeemkh Thank you for the suggestions. I've made the changes as requested. The latest paper is downloadable from [here](https://github.com/OpenDebates/openskill.py/actions/runs/7412597125/artifacts/1148149351).
