### 2024-01-04 549909a
Remove unnecessary bibtex references.
Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
