### 2024-01-05 #5901 @matt-graham
Apologies for the delay in getting back to this after the holiday period. I've now finished my review checklist based on the changes @vivekjoshy has made in https://github.com/OpenDebates/openskill.py/pull/116 - when that is merged the paper is ready for acceptance from my perspective.
