### 2024-01-06 #5901 @Naeemkh
@vivekjoshy, Thank you for the update. Please note that using hyperlinks as references is not considered appropriate in technical documentation. Instead, formal references should be used. Thanks.

For instance (or any other format):

```
@misc{busby_openskill.ex_2024,
  author = {Busby, Philihp},
  title = {{openskill.ex: Elixir implementation of Weng-Lin Bayesian ranking}},
  year = {2024},
  publisher = {GitHub},
  url = {https://github.com/philihp/openskill.ex},
  commit = {latest},
  note = {Last accessed on January 5, 2024}
}
```
