### 2024-01-06 #5901 @vivekjoshy
@Naeemkh There is no author information for the Lua implementation since only their GitHub username is available. How would I  cite them using bibtex?
