### 2024-01-06 8916610
Add latest suggestions by reviewers.
- Add ipywidgets to notebook.
- Rerun the notebooks.
- Update paper with latest benchmark results.
- Add bibtex references for OpenSkill ports.

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>
