+++ paper/paper.bib
@@ -146,3 +146,23 @@ year={2018}, month={Sep}
     url = {https://doi.org/10.1093/biomet/39.3-4.324},
     eprint = {https://academic.oup.com/biomet/article-pdf/39/3-4/324/930466/39-3-4-324.pdf},
 }
+@misc{PhilihpOpenSkillElixir,
+	author = {Philihp Busby},
+	title = {{G}it{H}ub - philihp/openskill.ex: {E}lixir implementation of {W}eng-{L}in {B}ayesian ranking, a better, license-free alternative to {T}rue{S}kill --- github.com},
+	howpublished = {\url{https://github.com/philihp/openskill.ex}},
+	year = {2020},
+	note = {[Accessed 04-01-2024]},
+}
+@misc{BrezinaOpenSkillKotlin,
+	author = {Jan Brezina},
+	title = {{G}it{H}ub - brezinajn/openskill.kt --- github.com},
+	howpublished = {\url{https://github.com/brezinajn/openskill.kt}},
+	year = {2022},
+	note = {[Accessed 04-01-2024]},
+}
+@misc{BstummerOpenSkillLua,
+	title = {{G}it{H}ub - bstummer/openskill.lua --- github.com},
+	howpublished = {\url{https://github.com/bstummer/openskill.lua}},
+	year = {2022},
+	note = {[Accessed 04-01-2024]},
+}
