### 2024-01-06 #5901 @Naeemkh
@vivekjoshy, I believe it can be regarded as a webpage. Here's an example:

```
@misc{openskill_lua_2024,
  title = {{openskill.lua}},
  year = {2024},
  url = {https://github.com/bstummer/openskill.lua},
  note = {Accessed: 2024-01-06}
}

```
