### 2024-01-06 #5901 @vivekjoshy
@Naeemkh @matt-graham If everything's fine shall I merge https://github.com/OpenDebates/openskill.py/pull/116 to main and try and generate the PDF using @editorialbot?
