### 2024-01-06 c7dac59
Update paper to include suggestions by reviewer (#116)
* Update paper to include suggestions by reviewer

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Add discussion section as suggested by reviewer

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Incorporate changes requested by Matt Graham

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Remove Git LFS

Signed-off-by: Vivek Joshy <8206808+vivekjoshy@users.noreply.github.com>

* Incroporate latest suggestion given by Naeemkh.

https://github.com/openjournal
[…]
