+++ paper/paper.bib
@@ -1,12 +1,12 @@
 @article{JMLR:v12:weng11a,
-  author  = {Ruby C. Weng and Chih-Jen Lin},
-  title   = {A Bayesian Approximation Method for Online Ranking},
-  journal = {Journal of Machine Learning Research},
-  year    = {2011},
-  volume  = {12},
-  number  = {9},
-  pages   = {267--300},
-  url     = {http://jmlr.org/papers/v12/weng11a.html}
+author  = {Ruby C. Weng and Chih-Jen Lin},
+title   = {A {B}ayesian Approximation Method for Online Ranking},
+journal = {Journal of Machine Learning Research},
+year    = {2011},
+volume  = {12},
+number  = {9},
+pages   = {267--300},
+url     = {http://jmlr.org/papers/v12/weng11a.html}
 }
 @article{Elo,
 title = {The psychology of esports players’ ELO Hell: Motivated bias in League of Legends and its impact on players’ overestimation of skill},
@@ -23,18 +23,146 @@ abstract = {This paper examines the folk theory of ELO Hell, which stems from th
 }
 @techreport{herbrich2006trueskill,
 author = {Herbrich, Ralf and Graepel, Thore},
-title = {TrueSkill(TM): A Bayesian Skill Rating System},
+title = {TrueSkill(TM): A {B}ayesian Skill Rating System},
 year = {2006},
 month = {January},
-abstract = {We present a new Bayesian skill rating system which can be viewed as a generalisation of the Elo system used in Chess. The new system tracks the uncertainty about player skills, explicitly models draws, can deal with any number of competing entities and can infer individual skills from team results. Inference is performed by approximate message passing on a factor graph representation of the model. We present experimental evidence on the increased accuracy and convergence speed of the system compared to Elo and report on our experience with the new rating system running in a large-scale commercial online gaming service under the name of TrueSkill.},
+abstract = {We present a new {B}ayesian skill rating system which can be viewed as a generalisation of the Elo system used in Chess. The new system tracks the uncertainty about player skills, explicitly models draws, can deal with any number of competing entities and can infer individual skills from team results. Inference is performed by approximate message passing on a factor graph representation of the model. We present experimental evidence on the increased accuracy and convergence speed of the system compared to Elo and report on our experience with the new rating system running in a large-scale commercial online gaming service under the name of TrueSkill.},
 url = {https://www.microsoft.com/en-us/research/publication/trueskilltm-a-bayesian-skill-rating-system-2/},
 number = {MSR-TR-2006-80},
 }
 @misc{suarez2019neural,
-      title={Neural MMO: A Massively Multiagent Game Environment for Training and Evaluating Intelligent Agents},
-      author={Joseph Suarez and Yilun Du and Phillip Isola and Igor Mordatch},
-      year={2019},
-      eprint={1903.00784},
-      archivePrefix={arXiv},
-      primaryClass={cs.MA}
+title={Neural MMO: A Massively Multiagent Game Environment for Training and Evaluating Intelligent Agents},
+author={Joseph Suarez and Yilun Du and Phillip Isola and Igor Mordatch},
+year={2019},
+eprint={1903.00784},
+archivePrefix={arXiv},
+primaryClass={cs.MA}
+}
+@misc{Rico_2022,
+title={Her ranking is a 10 but her skill rating says...},
+url={https://theanalyst.com/eu/2022/08/true-tennis-skill-ratings/},
+journal={The Analyst},
+publisher={Opta Analyst},
+author={Rico, Ysabel González},
+year={2022},
+month={Aug}
+}
+@article{10.1145/1852102.1852106,
[…]
