+++ benchmark/processors/__init__.py
@@ -1,4 +1,0 @@
-from .draw import Draw
-from .large import Large
-from .rank import Rank
-from .win import Win
