+++ benchmark/README.md
@@ -1,12 +1,11 @@
 # Benchmark Instructions
 
-Simply run ``benchmark.py`` with a compatible Python version and choose the options.
+Simply run the respective Jupyter Notebook files with a compatible Python version and choose the options.
 
 ## Available Benchmarks
 
-- ``Win``: Compares win performance against TrueSkill.
-- ``Draw``: Predicts draws on standard chess matches.
-- ``Rank``: Predicts the rank of players.
-- ``Large``: Uses rank prediction on a large multi-faction dataset.
+- ``win.ipynb``: Compares win performance against TrueSkill.
+- ``draw.ipynb``: Predicts draws on standard chess matches.
+- ``rank.ipynb``: Uses rank prediction on a large dataset.
 
-The ``Large`` benchmark requires the ``pubg.csv`` file (around 1 GB) be extracted from ``pubg.7z`` and place in the ``data`` folder.
+Note that the rank benchmark downloads a large file (around 1 GB) and requires slightly more RAM if running locally.
