+++ docs/doc_requires.txt
@@ -1,13 +1,13 @@
 sphinx~=7.0
 nbsphinx~=0.9
 pygments~=2.15
-shibuya~=2023.3
+shibuya~=2024.1
 ipykernel~=6.25
 myst_parser~=2.0  # Must Be Underscore, not Hyphen
 sphinx-intl~=2.1
 sphinx_favicon~=1.0
 sphinx-copybutton~=0.5
-sphinx-autoapi~=2.1.1
+sphinx-autoapi~=3.0.0
 sphinxext-opengraph~=0.8
-sphinxcontrib-bibtex~=2.5.0
+sphinxcontrib-bibtex~=2.6.1
 sphinx-autodoc-typehints~=1.25.0
