+++ pyproject.toml
@@ -81,15 +81,15 @@ docs = [
     "sphinx~=7.0",
     "nbsphinx~=0.9",
     "pygments~=2.15",
-    "shibuya~=2023.7",
+    "shibuya~=2024.1",
     "ipykernel~=6.25",
     "myst-parser~=2.0",
     "sphinx-intl~=2.1",
     "sphinx-favicon~=1.0",
     "sphinx-copybutton~=0.5",
-    "sphinx-autoapi~=2.1.1",
+    "sphinx-autoapi~=3.0.0",
     "sphinxext-opengraph~=0.8",
-    "sphinxcontrib-bibtex~=2.5.0",
+    "sphinxcontrib-bibtex~=2.6.1",
     "sphinx-autodoc-typehints~=1.24.0",
 ]
 tests = [
