### 2024-05-09 #5901 @vissarion
@vivekjoshy I do not know if you can do this after publication.
I do not know if this satisfies you but the link https://github.com/OpenDebates/openskill.py resolves to https://github.com/vivekjoshy/openskill.py
